refactor(api-credentials): split telemetry boundaries
